+++ metadata-ingestion/src/datahub/ingestion/source/looker/looker_common.py
@@ -784,6 +784,38 @@ class LookerExplore:
         return [mce, mcp]
 
 
+class LookerExploreRegistry:
+    """A caching registry of Looker Explores"""
+
+    def __init__(
+        self,
+        client: Looker31SDK,
+        report: SourceReport,
+        transport_options: Optional[TransportOptions],
+    ):
+        self.client = client
+        self.report = report
+        self.transport_options = transport_options
+        self.explore_cache: Dict[Tuple[str, str], Optional[LookerExplore]] = {}
+
+    def get_explore(self, model: str, explore: str) -> Optional[LookerExplore]:
+        if (model, explore) not in self.explore_cache:
+            looker_explore = LookerExplore.from_api(
+                model,
+                explore,
+                self.client,
+                self.report,
+                transport_options=self.transport_options,
+            )
+            self.explore_cache[(model, explore)] = looker_explore
+        return self.explore_cache[(model, explore)]
+
+    def get_all_explores(self) -> Iterable[LookerExplore]:
+        for key, value in self.explore_cache.items():
+            if value is not None:
+                yield value
+
+
 class StageLatency(Report):
     name: str
     start_time: Optional[datetime.datetime]
